Split AbstractFieldDirectiveResolver
